Remove unused -ac in the ffmpeg mp3 process
This ffmpeg command had two uses of -ac. The first would be ignored anyhow so I removed it. Also -ac is used to imply channel count, but the ignored -ac part here counted audio tracks and not channels. If the first -ac was used then single-track stereo audio files would have been converted to mono mp3s (so glad it was ignored).