Added flexible configuration management.
refs #4803

We previously only allowed for a default settings.py and a "local" version to
override the defaults. Changed things to use a more flexible system where
settings can be created that inherit from a "base.py" settings file. Once
created the settings can be specified in the django.wsgi file.