Updated manual normalization workflow for files with same name
When manually normalizing, if you have multiple files with the same
name, Archivematica doesn't know which manually normalized files go with
which original files.  This update allows a CSV file named
normalization.csv to be placed in the manuallyNormalized directory
specifying which access and preservation files correspond to which
origin files.

Update incomplete; works partially.

Issue #4922