Allow conflicting and not conflicting names for manual normalization
When providing a normalization.csv, allow files to be manually normalized
but not specified in CSV.

Eg.
image.bmp and image.gif are both manually normalized, use
normalization.csv to differentiate the manually normalized files.
text.doc is also manually normalized, and not specified in the CSV, and
still uses the manually normalized files