Display dashboard UUID and storage service URL in installer
Display pipeline UUID in installer, and give user the option to set the
storage service URL.  If accept defaults, set the storage service URL to
the one provided and create default Locations.  Otherwise, just register
the Pipeline (with no error if it fails), and open the storage service
in a new window.

refs #5259