AIP storage locations pulled immediately before chosen
refs #5440

Instead of Store AIP Location getting choices from a pre-built
MicroServiceChoiceReplacementDic, Archivematica queries the storage
service for a list of AIP storage locations and displays that to the
user.  A new MicroService was added to do the query.  If no storage
locations can be retrieved, goes back to the 'Store AIP?' MicroService.