Added FPR v2 way of identifying file format
refs #5575

Instead of running all the identification tools with FITS and selecting
which identification tool's output to use in the Ingest tab, we now run
just the identification tool selected, and no longer have a choice in
the ingest tab.

Client script identifyFileFormat added that runs the selected
script and compare the output against IDRules to determine the
FormatVersion of the file.  Removed the file identification portion of
archivematicaFITs client script.  Also removed the 'select format
identification tool' microservice and all of it's children.