archivematicaClient: coerce dashboardUUID to str
dashboardUUID may be returned from the database as a unicode instead
of an str. This causes an enormous amount of problems later when
performing the replacement of placeholder values with actual values.

The problem occurs if any of the arguments to a command to be run
contain unicode, or other non-ASCII, text. The text is brought in from
Gearman as cPickle data and is unmarshalled as an str, regardless of its
original content. Later on, the placeholder values are replaced using
the str replace() method using the values in the replacementDic dict
as arguments.

When python encounters string replacements where the object is an str
and one of the arguments is a unicode, or vice versa, python tries to
coerce both objects into the same type. Unfortuntaely when doing so, it
assumes that "str" is synonymous with "ASCII", and tries to convert the
str into a unicode as though it were ASCII. This fails if any character
values above 128 are present, which will of course be the case if the
value is actually UTF-8 being represented as an str.

fixes #5668