Better transfer backlog error handling.
refs #5681

When moving transfer files to the backlog directory, each
transferfile document in the transfers ElasticSearch index
gets its status updated. To find which files should be in
a transfer the Files MySQL table is consulted. If a file is
found in Files that doesn't have a corresponding transferfile
document, the script was trying to proceed as normal and
generating a stack trace.