File ID for DSPACE and Maildir transfers uses FPR v2
refs #5759

Merged part of the Standard, DSPACE and Maildir transfers so that they all
used the new FPR model for file format identification.  Since Maildir
needs to run file ID only on attachments, created new links for
maildir and used a Unit Variable to determine split.  Removed extra
MicroServiceChainLinks and TasksConfigs, along with other cleanup in the
Transfer area of the chains.

Enabled LoadMagicLink and UnitVariableLinkPull to pass along passVars.
Updated MCPGraphing for the changes in transfer.