File ID and Normalization works for maildir
refs #5759, #5248

Add MicroServiceChainLinks for maildir-specific file format identification
in Ingest, and added a check early in the SIP creation process to see if a
SIP came from a Maildir transfer.  If so, set UnitVariables for the correct
file ID chain link, and what directory to normalize.

Added a check in linkTaskManagerSplit to look for that unit variable and
use that directory for normalization if found.