Added createAICMETS client script
refs #5813

Added a client script to create the METS file for an AIC.  It collects AIP
information from files with an AIP uuid and contents of the AIP name, and
gets the AIP label by extracting the Dublin Core title from the AIP's METS
file.  Using that an the AIC's Dublin Core metadata, it creates the METS
file.

Updated archivematicaCommon/storageService.py to not rely on Django's ORM,
so all the client scripts that use it don't need to include Django's
settings file.  Added an extract file function to storageService that
fetches and saves a file from a Package (ie AIP)