Fixed METS file for AICs
refs #5883

Added metadata/ to the createMETS2 script, and added a check if the
directory exists before adding it to the structmap.  Also created a new
StandardTasksConfig for creating METS for an AIC that doesn't include the
amdSec.  Added the AIC METS file to Derivations, so it would work with
createMETS2.

Converted createAICMETS and createPointerFile to use
archivematicaXMLNamesSpace.  Renamed dctermsNS to dcNS, and removed unused
dcNS.