Add count of AIPs in AIC to ES and Archival Storage display
refs #5898

Want to be able to display the number of AIPs semantically included in an
AIC.  To do this, we index the number of AIPs in the AIC METS file into
ElasticSearch, and store that in the AIC METS file.

When creating the AIC METS file, we save the number of AIPS in a
UnitVariable for that AIC, and in the extent Dublin Core tag.  IndexAIP
looks for this variable if the SIP type is AIC and optionally passes it to
ElasticSearch, where it is retrieved by the Archival Storage tab.