Update Archivematica's version number
refs #6088

Add a version document in archivematicaCommon with the version and a
function to return a string to display, inspired by how Django does
versions.  Replace hard-coded values of Archivematica's version with a
call to the version function, including updating the version in the
installer.

Update MCP SQL dump to insert Archivematica-1.0.  Remove unused
createXmlEventsAssist.py