Decouple storage service helper from Django
Remove dependency between the storage service helper functions and
Django. Instead of using the dashboard helper's get_setting, use SQL
and databaseInterface to query for pipeline UUID.

This reduces problems when using in client scripts (that otherwise do
not need Django's settings), and allows the storage service to be used
in cases where the dashboard is difficult or impossible to import.