Get processing location from storage service
refs #6081

Archivematica should be more flexible on where its processing space is, and
let the storage service tell it where to put intermediate files.  To enable
this, create the shared directory structure programmatically when the
MCPServer starts up, once the currently processing location is know.

On boot, the MCPServer polls for a dashboard UUID and storage service
provided currently processing location.  Once this is known, create the
shared directory structure, set the owner and permissions, and update the
config files with this information.

Update installer scripts to accomodate this: /usr/share/archivematica is a
symlink to src/MCPServer/share in dev versions, cleaning watchedDirectories
removes all files in the shared directory, and postinst does not create the
shared directory structure anymore.  Clean up archivematicaMCP.

If pipeline UUID is None, get_pipeline also returns None.