Re-added FITS format ID info to METS
refs #6160, #6161

Re-added FITS format identification and associated event to the FITS
script.

When identification and characterization etc were split into separate
scripts, FITS' handling of the format identification and event to go
into the METS file was lost, so no format information for the
original file was recorded, and there was no format identification
event.  Until a better solution is found, revert to the old way of
doing it.