Update pre-METS creation metadata reminder
refs #6217, #5216

Both 5216 and 6217 require a wait before the AIP METS is created to
allow the user to enter metadata.  Removed the 5216 manually normalized
files only wait, and updated the more general wait to work in all cases.

Moved the wait to before the MicroServiceChainLink split and added a
watched directory, removed the manually normalized files only wait and
associated watched directory.