Update backlog files' status
refs #6131

Update arranged status of files to track when a SIP fails or completes
successfully.

Index AIP marks files an being in an AIP, and submits a
deletion request to the storage service when all files in a transfer
are in an AIP.  Transfer UUID is tracked for each arranged file to
enable this.

Failed and rejected SIPs run failedSIPCleanup.py, which deletes
SIPArrange rows for each file in the SIP, making them arrangeable again.