Remove duplicate chain post-normalization
Post-normalization, a full chain is repeated twice. The only difference
between the two chains is a variable set at the very end; otherwise
every chainlink is identical.

The duplicated chains were making it difficult to introduce new
microservices that occured in this part of the workflow, especially
when introducing new choices.

This commit unifies the two chains. They branch at the very beginning,
to set the unit variable, and then converge into an identical set of
chainlinks.