Add term administration pages
These pages allow the existing taxonomy and taxonomy terms to be
browsed and edited.

Note that, while these exist in the source and are enabled by going
directly to their URLs, they are not actually linked to from anywhere.
The feature is considered too specific to be generally useful right
now.

Consider this a secret feature! If you're a source-diver, you can find
the URLs and try it out.

refs #5900