FPRClient: support paging API responses
Rather than try to fetch all items at once (which doesn't work for
sets of results larger than 1000), rework the existing code into a
simple fetch function and a generator that iterates over every item
of the requested type. These are now fetched in batches of 100 and
the "next" URL provided by the API response is used to page through
all results.

Fixes #6964.