csv: use universal newlines
This fixes reading CSVs with unusual newlines, like \r. We should use
this in all places that we read CSVs, in case they were produced by
systems that use those newlines.

FileMaker, for example, still outputs CSV with \r to this day, even
though MacOS transitioned to \n in 2001.

Fixes #7015.