Enforce valid user interaction in arrange panel.
Added disabling of SIP arrange panel action buttons when the user has
highlighted an item that the action should not apply to (creating
directories within files, for example).

Added an error alert in the SIP arrange UI if someone tries to
create a directory within a file.

Commented and cleaned up JavaScript.

Conflicts:
	src/dashboard/src/media/js/ingest/ingest_file_browser.js