MCPServer sets currentPath in watchedDirectories
When starting a chain from a watched directory, ensure a SIP's
current path matches the one found in the directory.  This can happen
during AIP reingest, where the last know location is not the watched
directory it is currently in. Do not change currentPath for DIPs, as
they can run in parallel with the originating SIP and overwrite the
SIP's information.