removing some delete statements from upgrade script
refs #7249
MicroServiceChainLinks cannot be deleted from an upgraded system
as there may be existing jobs that have foreign key links to those
MCSL's.

Rather than delete some MCSL's from an upgraded database, it seems
better to keep all upgraded systems in the same state.

This results in some orphaned MicroServiceChains, that do not
appear in a fresh install, but these orphaned Chains will not
be reachable, they will exist only to satisfy the foreign key
constraints (and provide context for the historical jobs data).