connect_and_remove_transfer_files: delete full transfer from ES
Previously, this function queried the database for a list of files in
the transfer, and removed those specifically. However, since the
search index contains many files that aren't in the database (such
as logs and metadata files), this left dangling files that could
never be removed.

Instead query the database for transfers, then delete all files in ES
from each of those transfers.

refs #7357