indexAIP: index MODS identifiers
This extracts identifiers from MODS documents contained within an AIP's
submissionDocumentation, and indexes them in the search index.

A new key has been added to the ES JSON document: 'identifiers'. This
key contains an array of zero or more strings, which are intended to
be identifiers from other systems cataloguing this data. For example,
in a transfer from Islandora, the identifiers might represent the
original identifiers of items in Islandora.

These MODS documents are looked up within the
submissionDocumentation/mods directories from transfers within the AIP;
they should be present within the submissionDocumentation/mods
directory in the transfer at the time that the transfer begins.

To support this, submissionDocumentation is copied back into the SIP
in order to ensure it's not deleted with the rest of the bagged SIP.

refs #7424