createMETS2: move option parser into __main__
With the option parser in the main script body,
it becomes unsafe to import anything from this
script unless the contents of ARGV are exactly as
expected by createMETS2’s parser.

Note that this produces a number of SyntaxWarnings,
and it is still unsafe to use certain functions if
one of these global variables is not defined.
Further cleanup should make these globals into
function parameters.