Dashboard: Accession IDs associate with Transfers again
refs #7442

For packages, we still want to add the package name to the transfer path.
Only append a / if it is a directory. Otherwise, a Transfer is created in
the dashboard that does not match the path of the transfer on disk, so
MCPClient creates another Transfer when it sees the new file.

Also, ensure files and folders have unique names by padding with numbers.