CONTENTdm: Remove Direct Upload
Remove direct upload as an option from CONTENTdm DIPs. All CDM DIPs are
project client.

No communication with the CDM server is necessary now, so remove MSCLs and
configuration options that deal with that. Remove code in restructure DIP
that deals with direct upload.

Output simple/compound.txt to the objects directory of the DIP.