MCPClient: File to Folder client script
Add client script that checks if a SIP is a file, and moves it into an
identically named folder if that is the case.  Allows transfer types to
handle both files and folders as input.

Move transfer no longer assumes all transfers are directories and only
appends a / if the source is a directory.