sanitize: Don't write events for directory name sanitization
These events were associated with files (not directories), were
repeated several times over in the METS file, and were highly
misleading.

We don't currently treat directories as intellectual entities for the
sake of PREMIS, but this might change in the future. Meanwhile, the
sanitization log file (which is stored in the AIP) has accurate
information on the sanitization process and should be preferred for
looking up directory names.

Fixes #8033.