Archival Storage: Pass full results to create AIC
refs #8258

CreateAICForm should be passed the full list of AIPs that match the query,
not just this page's results. The list needs to contain the AIP name and
UUID. Modify the query to return those fields and format the data as
expected, to reduce the number of calls to search_augment_aip_results.

Remove facets from assemble_query since they're no longer used. Add sorting
to request for consistent results.