taskStandard: stdout/stderr values are already normalized
Due to the changes to ReplacementDict, the stdout/stderr values are
now being expanded before being passed to taskStandard.

Fortunately, this allows for a general simplification/cleanup of the
validation code in taskStandard; it can now perform a straightforward
check whether the files are writable, instead of relying on heuristics
based on the presence of a variable string in the text.

Fixes #8404.