Reingest: Rights & DC marked with reingest status
refs #6488

DC & PREMIS rights metadata should be written to the METS file only if it has
been changed during reingest. To track this, add a status field to the
DublinCore & RightsStatement tables which can be ORIGINAL, REINGEST or
UPDATED. ORIGINAL is the default for AIPs being created normally, REINGEST is
the status when it is initially parsed to the DB, UPDATED is when the data is
changed after reingest.

Fix DublinCore.metadataappliestotype to be a foreign key.