Forms: normalize and validate dates
These were previously treated as pure text, so any value was permissible.
However, Elasticsearch may now opportunistically parse these into dates,
which means that we have to be more careful about validating them.
Use dateutil to try to normalize these values; if we can't, raise a
ValidationError to provide the user the chance to fix it.

Fixes #8788.