Correcting duplicate files checksums verification.
This commit changes the behaviour of the transfer checksums verification to
check correctly duplicate files (same content, different file names).

The microservice Verify transfer checksums (Job: Verify metadata directory
checksums) does not work properly in case there are same files with different
names in the transfer. For example (hash is md5 but it is the same for sha1 and sha256):

89808652ad47f2bb97df733e8a2e14ba objects/collection/folder/a.jpg
89808652ad47f2bb97df733e8a2e14ba objects/collection/folder/b.jpg

Such transfer fails and the a.jpg file is reported as corrupted (hash does not
match) because the md5deep is called

md5deep -r -n -m

which compares the computed hashes first (so the computed hash for the b.jpg
matches the line for a.jpg).

This behaviour catches duplicate files which is probably not the goal of the
Verify transfer checksums microservice.

This request call for the <hash>deep should be

md5deep -r -x

This slight change in the call works as intended: reports only those files not
listed in checksum.<hash> or those with corrupted (mismatched) hashes.
Duplicate files are matched correctly (and not reported).