Correct dump file names
Filenames src/MCPServer/share/1:1.4.0 and src/MCPServer/share/1:1.4.1 seems to
be breaking dump files naming conventions and were renamed to
src/MCPServer/share/1.4.0 and src/MCPServer/share/1.4.1 respectively.

Moreover, colons (‘:’) in filenames are prohibited characters on Windows and as
such it is not possible to fully checkout working copy on Windows filesystem.