Enable use of repository in ASpace config
refs #9075

The ArchivesSpace DIP Upload settings page,
in the admistration tab of the dashboard,
allows a user to set a value for the repository
to connect to in ArchivesSpace.

This value is now saved to the ReplacementDic
and used in the matching gui.

Also adding support for changing the repo number in
dip_generation_helper.py