Archival storage: fix AIP pagination
This reworks archival storage's search view in order to fix paging of
AIP searches.

AIP searches are very broken in the current iteration. For historical
reasons, searches are performed against the aips/aipfile index, which
is limited to returning only AIP-related fields. This is done because
aips/aipfile contains relevant information which is not present in the
aips/aip index, which contains actual AIP data.

The issue arises when paging is applied. Because these are file results,
there are many files to one AIP; these are incorrectly paged as file
results. This means that each page, instead of containing ten AIPs as
intended, contains ten files' worth of AIPs - which could be a single
AIP, if the first returned AIP contains more than one file.

In addition to fetching metadata from aips/aipfile, the view later uses
the fetched UUIDs to fetch records from aips/aip - in order to have
enough information to seed the AIC form, and some other additional
metadata that aips/aipfile is missing. It currently does this by
performing one query for every fetched UUID, which is fairly wasteful.

This patch simplifies the current workflow to the following:

* Run the query against aips/aipfile, with and an aggregation which
returns all unique UUIDs in the requested set and the count of files
within that UUID. This query is set to return no results - only the
aggregation is used.
* Uses that set of UUIDs to fetch pages of AIP data from aips/aip in a
single query, which feeds the template.

Because the paging is now based on AIPs, not files within an AIP, a
page in archival storage can display the appropriate number of AIPs.
In addition, because AIPs are fetched in pages and the total number of
queries performed has been reduced substantially, performance is
substantially better.

Fixes #9030.
Fixes #9085.