METS: Update for Event/Agents many-to-many relationship
Events only link to Agents explicitly associated with them. Agents are only
added to an amdSec if they're associated with Events in that amdSec.

createDigivProvMD creates both Events and Agents, as described. Created a new
createAgent that mimics createEvent in functionality. Update callers.

Add & update tests, update fixtures to be more modular.