MCPClient: use thread-safe logger
One of the many benefits of `logging` is its inherent thread safety. Using it
saves us from having to deal with thread synchronization in order to provide our
worker threads with safe access to `stdout` and `stderr`. This commits deletes
`printOutputLock` and replaces the old print statements with calls to the
logger.