elasticSearchFunctions: receive conn parameters
This commits refactors `elasticSearchFunctions` so the connection settings
are *received* from its user instead of being looked up in `clientConf`. This
module is used by the `dashboard` and `MCPClient` (in particular, client
scripts), being `clientConf` available only in the former case.

The goal is to enable the dashboard to run separately from the MCPClient.