django.wsgi: add string environment support
Set process environment variables on each request based on the per request WSGI
environ settings set using SetEnv in Apache.

Enable monitor when `settings.DEBUG` equals True. Because mod_wsgi is running
in daemon mode, the only way to populate the changes is to touch the
`django.wsgi` file. The `monitor` module will be responsible for that from a
separate daemonic thread.

DJANGO_SETTINGS_MODULE is expected to be defined in Apache. A default value is
not provided (you may expect `settings.common`, but I think it's better not to
make assumptions about your environment).

This is how the configuration could look like in Apache:

    SetEnv DJANGO_SETTINGS_MODULE settings.local
    SetEnv ARCHIVEMATICA_DASHBOARD_FOO bar
    WSGIScriptAlias / /usr/share/archivematica/dashboard/apache/django.wsgi
    WSGIDaemonProcess dashboard user=archivematicadashboard group=archivematica python-path=/usr/share/archivematica/dashboard:/usr/lib/archivematica/archivematicaCommon
    WSGIProcessGroup dashboard

Note that PYTHONPATH is passed to `WSGIDaemonProcess`.