elasticSearchFunctions: global client as module attribute
Elasticsearch's client is thread-safe. We use it in two different contents:
our `dashboard` application and `clientScripts`.

This commit stores the Elasticserach instance as an attribute of its module.
Access to the instance if provided via `get_client()`, but it requires previous
initialization. In the `dashboard this is achieved in `apache/django.wsgi` while
clientScripts have to initialize it manually for now. An exception is raised if
the user tries to access to the client before initialization.

Usage example:

    import elasticSearchFunctions
    elasticSearchFunctions.setup(hosts=['127.0.0.1:9200'])
    client = elasticSearchFunctions.get_client()
    results = client.search(**{})

It's been provided an extra setup method that reads the configuration param
from `clientConf.conf` to ease the transaction. This will be used by both
`dashboard` and `clientScripts` until they enable other ways to provide that
configuration, which is work in progress.

In addition, the dashboard receives a new middleware that replaces the old
`elasticsearch_required`. See `middleware.common.ElasticsearchMiddleware`. It
redirects the user to a custom Elasticsearch error page when one of the two
following exceptions have been raised in the view: `ElasticsearchException` and
`ImproperlyConfigured`. The main advantage here is to avoid an extra hit to the
Elasticsearch server (as the previous `check_server_status` function used to
do).