MCPServer: wait until database is available
I'm trying to eliminate the need to deploy services in specific order. This
commit allows us to start MCPServer even when MySQL is not available. But the
application will block and wait during initialization until it is available.

Stanza `start` in Upstart service has been update to stop depending on a local
MySQL service.