Index transfer contents once
The current indexing process indexes files near the end of the
transfer process, before a storage decision has been made.
Subsequently, if the user decides to send the files to backlog, the
records must be retrieved from the index and updated with a new status.

This instead defers indexing until after a storage decision has been
made, and indexes them once with the final status.

Files which are not sent to the backlog are no longer indexed; those
records were never used.

Fixes #9265.