Revert "ArchivesSpace: Do not open matching GUI with JS"
This reverts commit c2cc8f6cee06fdb79c0851d8a607837f19ee8c07.

After some discussion, we determined that this was not the best solution to
the problem at this point in time. We don't have a clear UI that drives users
to this feature, so expecting users to know to click on the small clipboard
icon seems liable to result in a great deal of confusion.

This shouldn't affect most automation users, as Holly suggests they may
well not notice the popup is there or even have the dashboard UI open
at the point at which it would be triggered.

Fixes #9347.