dashboard: join host:port str gearman setting
That will simplify later the management of gearman servers with environment
variables, e.g. we could pass a string like "host1:port1;host2:port2" but
breaking host and port in two different variables would make that harder to
manage.

Also, renamed the setting from `MCP_SERVER` to `GEARMAN_SERVER` to be more
clear about the host that we are consuming in reality.