Don’t autodetect METS mappings
In the past we used autodetection for most METS mapping fields, but this
led to significant issues with type autodetection.

We tried to handle this in the past by special-casing certain fields
which might get autodetected as dates; however, that game of whack-a-
mole is going to be pretty hard to keep up. Instead, I’ve generated METS
mappings using as complete a METS file as I could and included those in
the archivematicaCommon tree as JSON files; when creating an index, we
now use those to specify the entire format at once.

Fixes #9330.