adding default values for config parameters.
WIP - this is an example only.  This is a way to make sure there
are default values for a development/testing environment.
The assumption is that the user will override these with their own
settings, in a config file, but if the config file is missing or
is missing required elements, the defaults defined here will be used.

Suggesting this in response to
https://github.com/artefactual-labs/ansible-role-archivematica-src/pull/53

When deploying Archivematica using that ansible role, a new config file is
used, that is not compatible with this branch of the Archivematica repo.

By specifying default values, A dev install can be completed using ansible.