Dashboard: strip SS API key field, refs #9838
Not entirely sure if this is the most idiomatic way to achieve it. One
limitation I've found is that the HTML document that Django returns after the
form is posted by the browser does not contain the stripped value but the
original, however the value stored is the cleaned one. I'm not sure if the
problem is in the way that the view is built or somewhere else.