sanitizeNames: Run on files, not on sanitization events
Previous behaviour generated name cleanup events based on which files had
been sanitized. This missed files that were in a sanitized directory but that
didn't need filename sanitization.

This iterates through all files in the unit and checks for sanitization on
all directories in the path.  If found, it generates a single sanitization
event from the totally unsanitized path to the completely sanitized path.