fixes issue #10759
For some reason (unknown to me), the value of input.filter(':text').val()
is empty when it is used to populate the form POST data sent to Archivematica.

This commit saves the value to a variable and uses the variable instead.

I'd love to know why this is necessary.  The code here did not change from
the 1.5.1 release, other than the change to a new bootstrap modal.

I assume that is the reason the behaviour has changed, but I am not sure
why it didn't show up until now.