Fixes #10785: AIC creation failing at restrForComp
This commit fixes a regression introduced by
07ca4925560afeb09753b3cc9065c8f7521da0b9. That commit assumed that every
UUID passed into restructureForCompliance.py would match a Transfer
model. However, when an AIC is created, the UUID corresponds to a SIP
model. With the present commit, AIC creation should work as it did in
stable/1.5. However, future work will need to refactor AIC creation since it
results in a (non-fatal) failure at "Job: Removed bagged files" (and it
creates directories (e.g., objects/ logs/) that are not needed for an
AIC.