Migrate Events with no linking agent
Refs #10831

The linkingAgentIndentifier column of the Events table is dropped
in migration 0026, being replaced with a many to many linking table
between Events and Agents.  When migrating existing data, it is possible
for the data from Archivematica 1.5.1 and earlier to have rows in the Events
table with no linkingAgentIdentifier, as the column is nullable.

Not sure how the rows get there, I would think that these are orphan rows,
probably a bug.  But they break the migration, so this change is to ignore
the rows with no linking agent when migrating to the new data model.