Deleted derivatives retained in AIP METS
When a reingest results in deleted derivatives and a subsequent reingest
is performed, the relevant
mets:fileSec/mets/fileGrp[@USE="deleted"]/mets:file elements are
retained as are the deletion digiprovMD elements.
NOTE: this is a WIP proof-of-concept commit which monkey-patches
mets-reader-writer; these changes should ultimately be merged to metsrw
itself.