Allow multiple originals for single access
When a user creates a transfer with objects/ and objects/access/
directories, allow for multiple files in objects/ to correspond (via
filename) to a single file in objects/access/. This prevents the "bug"
during checkForAccessDirectory_v0.0 encountered in #10979. However, it
is unclear if this is really a bug, i.e., if we want to support
transfers structured in this way.