Normalization: Use default thumbnail rule if normal rule errors
The error-then-default behaviour for access derivatives needed to be
ported to thumbnails also. If not, then a transfer containing a single
file that errors on thumbnail creation (with no thumbnail created) will
cause a failed SIP at Prepare DIP: Copy thumbnails to DIP directory
because the thumbnails/ dir will be empty and will have been removed,
causing `cp` to error. Refs #10779.