Dashboard: use WhiteNoise to serve static files
Django prefers not to be used to serve static files. Normally this is done from
Nginx which adds complexity to the deployment process. Particularly when using
containers this became a problem because the Nginx service does not have access
to the static assets of the web application.

Whitenoise enables Django to be deployed as a self-contained unit.

The changes under `src/media` are there only to fix broken reference to assets.
less.js is deleted because it seems unused.