MCPClient: stop messing with env
Removing hard-coded PYTHONPATH and using `settings.clientscripts` which is
a symbolic link to Dashboard's `settings.common`. Client scripts need to use
Dashboard settings instead of MCPClient settings. Using `settings.common` was
a problem because it was matching the latter first based on the order that
the imports appeared in PYTHONPATH.