Delete base64 git submodule
In practice we're using src/dashboard/src/media/js/vendor/base64 which was
never replaced with a symlink to the submodule. As we prefer to avoid git
submodules this commit just deletes it. If we need a more recent version of
base64.js in the future we'll just vendor it again or approach the dependency
differently, e.g. bower...