Dashboard: make IP whitelist optional
Before, you needed to have the user authorized *AND* a whitelisted IP address
which makes really hard or impossible to work with deployment environmnets like
Amazon ECS or similar where the origin IP address is not known in advance.

The `api_whitelist` setting is now set to empty string by default, which means
no whitelist is applied. This can be changed at install time by the
installer command.