Dashboard: require auth OR valid IP (instead of AND)
Before, you needed to have the user authorized *AND* a whitelisted IP address
which makes really hard or impossible to work with deployment environmnets like
Amazon ECS or similar where the origin IP address is not known in advance.

The `api_whitelist` setting is reset to empty string for security reasons.