TRIM: update to use Django models
refs #11161

The TRIM transfer type is normally not available in an Archivematica
dashboard.  If that option is enabled, a workflow is triggered that
uses these four scripts.

These scripts have not been updated since before the 1.4.0 release.
They call getFileUUIDLike(), a method that is defined in
archivematicaCommon/fileOperations.py, and that was updated in
version 1.4.0 to use Django Models.

As a result, the list of permissible values for the 4th parameter
changed.

This commit updates the TRIM workflow to call getFileUUIDLike
correctly.