Allow 'object count' from SS to be string
Fixes #636. If a transfer source location contains over 5000 files, then
filesystem_ajax/views.py will trigger a TypeError because the call to
ungettext assumes that the value of the 'object count' attribute
returned by the SS will always be an integer. In fact, in this case the
value will be a string, i.e., '5000+'. This change catches the
``TypeError`` and uses ``ugettext`` with ``%(...)s``.