MCPClient error check Gearman worker creation
This patch adds a try/except block to the MCPClient when creating
a Gearman worker in startThread().

Without this patch, if the MCPClient configuration item
"MCPArchivematicaServer" has an invalid value, no Gearman worker
will be created and Archivematica will be stuck thinking that a
job is executing indefinitely with no indication of what happened
in the user interface or the logs.

To test, open "/etc/archivematica/MCPClient/clientConfig.conf",
and change "MCPArchivematicaServer" to something invalid like
"buffalo" or "localhost::9999", and then try to do a standard
transfer in the Archivematica dashboard UI. In the micro-service
"Verify transfer compliance", you'll get stuck at "Job: Set file
permissions". It will say it's still executing but the job will
never actually run.