Added "Store DIP" choice in config page
This required also modifying the workflow (hence the migration) because
the post-DIP-upload "Store DIP?" choice link was being activated AFTER
relocation of the DIP to the uploadedDIPs/ directory, which resulted in
the MCPServer not knowing where the DIP was when it went to look for the
processingMCP.xml config file. Now moving the DIP to uploadedDIPs/ only
occurs after the upload/store DIP steps have completed.