MCPClient: antivirus pass by ref/val opt
In a36a9bf216 we added the ability to talk to a clamd instance over the
network but in addition to that we started passing the files to clamd by
reference which doesn't always work. This commit adds a new option to MCPClient
so the user can decide if the file should be passed by value or by reference.
Passing by value (streaming) is the default because that's the easiest setup.