MCPClient: remove createPointerFile.py MS
Introduces a migration that removes the "Create AIP Pointer File"
micro-service. Pointer file creation will be moved to the Storage Service (SS).
The motivation for this is that pointer files need to be created in the SS for
a different purpose, namely for the replica packages that are created when an
AIP Storage location has a mirror location; the replicas need their own pointer
files with replica-specific information.